Refactor admin files from JS to TS
