+++ admin/src/hooks/index.ts
@@ -1,3 +1,3 @@
 // Remove this eslint disable line once more then one hook exists.
 // eslint-disable-next-line import/prefer-default-export
-export { default as usePluginConfig } from './use-plugin-config';
+export { default as usePluginConfig } from './usePluginConfig';
