+++ admin/src/hooks/usePluginConfig.ts
@@ -1,16 +1,22 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useAPIErrorHandler, useFetchClient, useNotification } from '@strapi/helper-plugin';
+import { useAPIErrorHandler, useFetchClient, useNotification } from '@strapi/strapi/admin';
 
+import { type SnippetsPluginConfig } from '../../../server/src/config';
 import { ACTION_RESOLVE_CONFIG } from '../constants';
 import { pluginId } from '../utils';
 
-const usePluginConfig = () => {
+type UsePluginConfigReturn = {
+  data: SnippetsPluginConfig;
+  isLoading: boolean;
+};
+
+const usePluginConfig = (): UsePluginConfigReturn => {
   const { formatAPIError } = useAPIErrorHandler();
   const dispatch = useDispatch();
   const fetchClient = useFetchClient();
-  const toggleNotification = useNotification();
-  const { config, isLoading } = useSelector((state) => state[`${pluginId}_config`]);
+  const { toggleNotification } = useNotification();
+  const { config, isLoading } = useSelector((state: any) => state[`${pluginId}_config`]);
 
   useEffect(() => {
     // Do nothing if we have already loaded the data.
@@ -20,13 +26,13 @@ const usePluginConfig = () => {
 
     fetchClient
       .get(`/${pluginId}/config`)
-      .then(({ data }) => {
+      .then(({ data }: any) => {
         dispatch({
           type: ACTION_RESOLVE_CONFIG,
           data,
         });
       })
-      .catch((err) => {
+      .catch((err: any) => {
         if ('code' in err && err?.code === 'ERR_CANCELED') {
           return;
         }
@@ -36,7 +42,7 @@ const usePluginConfig = () => {
           message: formatAPIError(err),
         });
       });
-  }, [isLoading, config, dispatch, fetchClient, toggleNotification, formatAPIError]);
+  }, [isLoading, config, dispatch, fetchClient, formatAPIError, toggleNotification]);
 
   return {
     data: config,
