--- admin/src/utils/index.js
@@ -1,3 +1,0 @@
-export { default as pluginId } from './plugin-id';
-export { default as pluginName } from './plugin-name';
-export { default as sanitizeCode } from './sanitize-code';
