+++ admin/src/utils/index.ts
@@ -1,0 +1,3 @@
+export { default as pluginId } from './pluginId';
+export { default as pluginName } from './pluginName';
+export { default as sanitizeCode } from './sanitizeCode';
